[dtensor] refactor dispatch and add cached propagator
[ghstack-poisoned]
