+++ torch/distributed/_tensor/ops/math_ops.py
@@ -2,7 +2,7 @@
 from typing import cast, Optional, Sequence
 
 from torch.distributed._tensor.api import DTensor
-from torch.distributed._tensor.dispatch import OpSchema, OutputSharding
+from torch.distributed._tensor.prop import OpSchema, OutputSharding
 from torch.distributed._tensor.ops.common_rules import pointwise_rule, reduction_rule
 from torch.distributed._tensor.ops.utils import (
     as_list,
@@ -46,7 +46,7 @@ sum_ops = [
     "aten.sum.dim_IntList",
 ]
 for sum_op in sum_ops:
-    DTensor._op_to_rules[sum_op] = sum_rule
+    DTensor._propagator.register_sharding_prop_rule(sum_op, sum_rule)
 
 
 @register_prop_rule("aten._softmax.default")
@@ -96,7 +96,7 @@ mean_ops = [
 ]
 
 for mean_op in mean_ops:
-    DTensor._op_to_rules[mean_op] = mean_rule
+    DTensor._propagator.register_sharding_prop_rule(mean_op, mean_rule)
 
 
 def var_rule(op_schema: OpSchema) -> OutputSharding:
@@ -122,7 +122,7 @@ var_ops = [
 ]
 
 for var_op in var_ops:
-    DTensor._op_to_rules[var_op] = var_rule
+    DTensor._propagator.register_sharding_prop_rule(var_op, var_rule)
 
 
 @register_prop_rule("aten.var.correction")
