+++ torch/distributed/_tensor/ops/pointwise_ops.py
@@ -2,7 +2,7 @@
 from typing import cast
 
 from torch.distributed._tensor.api import DTensor
-from torch.distributed._tensor.dispatch import OpSchema, OutputSharding
+from torch.distributed._tensor.prop import OpSchema, OutputSharding
 from torch.distributed._tensor.ops.common_rules import (
     linear_pointwise_rule,
     pointwise_rule,
@@ -369,11 +369,11 @@ pointwise_ops = [
 
 
 for op in linear_pointwise_ops:
-    DTensor._op_to_rules[op] = linear_pointwise_rule
+    DTensor._propagator.register_sharding_prop_rule(op, linear_pointwise_rule)
 
 
 for op in pointwise_ops:
-    DTensor._op_to_rules[op] = pointwise_rule
+    DTensor._propagator.register_sharding_prop_rule(op, pointwise_rule)
 
 
 @register_prop_rule("aten.native_dropout.default")
