+++ torch/distributed/_tensor/prop.py
@@ -1,0 +1,155 @@
+from dataclasses import dataclass
+from typing import Callable, Dict, List, Optional, Tuple, Union, Tuple, Sequence
+
+import torch
+
+from torch.distributed._tensor.placement_types import DTensorSpec
+
+
+ArgKwargsType = Union[Tuple[object, ...], Dict[str, object]]
+# ATen op schemas could have Tensor, Tuple[Tensor] and List[Tensor], so output type sould
+# be the same set of possiblities.
+OutputSpecType = Optional[Union[DTensorSpec, Sequence[Optional[DTensorSpec]]]]
+
+
+# @dataclass(eq=True, frozen=True)
+@dataclass
+class OpSchema(object):
+    """
+    OpSchema is a data class that describes an operator input schemas, it
+    includes DTensor DTensorSpecs and non-tensor args/kwargs (positional order
+    preserved). It is mainly used by the dispatching logic below to run things like
+    sharding propagation.
+
+    Sharding propagation rules registered could utilize this data class and
+    do inplace update some fields (when necessary, i.e shape related ops) to make
+    sure the args/kwargs are legit before passing to the local tensor operator.
+    This is the main reason that we don't freeze this dataclass.
+
+    NOTE: greater access to the operator inputs comes with greater responsibility.
+    Here are some basic rules about what can be used and what can be changed.
+
+    Args:
+        func_schema: the function schema of the operator
+        args_schema: contains args except that the DTensor args have been replaced
+            with its DTensorSpec
+        kwargs_schema: contains kwargs except that the DTensor kwargs have been replaced
+            with its DTensorSpec
+
+    What can be used:
+        - every attribute within this class could be read to conduct
+          sharding propagation.
+    What can be changed:
+        - only the args_schema and kwargs_schema could be changed.
+        - every non-tensor args could be changed to accomodate for local tensor
+          operations (i.e. for ops like view/reshape/...)
+        - every "DTensorSpec" attribute inside `args_schema`, `kwargs_schema` and
+          `args_spec` SHOULD NOT be updated! DTensorSpec are read only and sharding
+          propagation shouldn't inplace update them, otherwise the input DTensor
+          placements will get implicitly changed and it's error-prone.
+    """
+
+    func_schema: torch._C.FunctionSchema
+    args_schema: Tuple[object, ...]
+    kwargs_schema: Dict[str, object]
+
+    @property
+    def args_spec(self) -> Tuple[DTensorSpec, ...]:
+        """
+        args_spec: Tuple[DTensorSpec, ...]: contains a clean list of args spec list
+            with NO non-DTensor positional arguments (i.e. int/float/tuple, etc)
+            mainly used by sharding propagation to propagate the output spec
+        """
+        # filter out non-relavant values from args schema to get a clean spec list
+        # this would mainly be used by sharding propagation rules
+        return tuple(item for item in self.args_schema if isinstance(item, DTensorSpec))
+
+    @property
+    def is_inplace(self) -> bool:
+        return self.func_schema.name[-1] == "_"
+
+    @property
+    def is_out_variant(self) -> bool:
+        return "out" in self.func_schema.overload_name
+
+    def __repr__(self) -> str:
+        return (
+            f"OpSchema(func_schema={self.func_schema},"
+            f" args_schema={self.args_schema},"
+            f" kwargs_schema={self.kwargs_schema})"
+        )
+
+    def __hash__(self) -> int:
+        repr_str = str(self.func_schema) + str(self.args_schema)
+        return hash(repr_str)
+    
+    def __eq__(self, __o: object) -> bool:
+        return hash(self) == hash(__o)
+        # repr_str = str(self.func_schema) + str(self.args_schema)
+
+        # pass
+
+
+@dataclass
+class OutputSharding:
+    """
+    OutputSharding is a data class that is used by the sharding propagation
+    rules, it could set the output_spec upon successful propagation, and if
+    it failed, output_spec would become None and sharding propagation rules
+    could give a list of suggestions for inputs to reshard.
+
+    NOTE: the schema_suggestion generated by sharding propagation should be
+    exactly the same as the operator OpSchema, except the DTensor DTensorSpecs
+    """
+
+    output_spec: OutputSpecType
+    schema_suggestions: Optional[List[OpSchema]] = None
+    failed_reason: Optional[str] = None
+
+
+class ShardingPropagator(object):
+    def __init__(self) -> None:
+        self.op_to_rules: Dict[str, Callable[[OpSchema]], OutputSharding] = {}
+
+    def register_sharding_prop_rule(self, op_key, rule_func):
+        self.op_to_rules[op_key] = rule_func
+
+    def propagate_op_sharding(self, op_overload, op_schema: OpSchema) -> OutputSharding:
+        op_key = str(op_overload)
+        sharding_prop_func = self.op_to_rules.get(op_key, None)
+
+        if sharding_prop_func is None:
+            # step 1. If there's not even one sharding rule
+            # implemented for the operator, we fall back to
+            # local tensor compute, this is wront currently
+            # we will change the behavior to reshard to full
+            # replicate and do the computatation
+            raise NotImplementedError(
+                f"Operator {op_key} does not have a DistributedTensor rule registered."
+            )
+
+        # step 2. there's sharding propagation rule, run
+        # sharding propagation to get output sharding
+        try:
+            output_sharding = sharding_prop_func(op_schema)
+        except Exception as e:
+            raise RuntimeError(
+                f"Sharding propagation failed on op {op_key}.\n"
+                f"Input schema: {op_schema}.\n"
+                f"Error: {e}"
+            ) from e
+
+        # step 3. if can't get output_spec from sharding
+        # propagation (i.e. no rules apply for input
+        # placements), we return the output sharding
+        # with schema suggestions, which can be used to
+        # decide how to do redistribute on inputs
+        if output_sharding.output_spec is None and \
+            output_sharding.schema_suggestions is None:
+                raise RuntimeError(
+                    f"Sharding propagation failed on op {op_key}!"
+                    f"Input schema: {op_schema}."
+                    f"Failed reason: {output_sharding.failed_reason}"
+                )
+        else:
+            return output_sharding
